Revert "Skip steps which do not require user interaction"
