+++ app/public/html/choose-account.twig
@@ -5,7 +5,7 @@
 {% block content %}
     <h1>Choose an account to import from and to</h1>
 
-    <form name="next-form" action="." method="post">
+    <form action="." method="post">
         <input type="hidden" name="step" value="{{ next_step }}">
         <div class="form-group">
             <label for="bank_account">Bank Account</label>
